refactor: move all functions to /filter
