+++ src/lib/filters/add-hash.js
@@ -1,12 +1,11 @@
-module.exports = function addHash(data, withHash) {
-  return new Promise((resolve, reject) => {
-    const withHash = data.map((item) => {
-      if (item.startsWith('#')) {
-        return item
-      } else {
-        return '#'.concat(item);
-      }
-    })
-    resolve(withHash)
+module.exports = function addHash(data, withHash) { // add hash to answers that don't start with a hash yet
+  return data.map((person) => {
+    let eyeColor = person.eyeColor
+    if (eyeColor.startsWith('#')) {
+      return {eyeColor: eyeColor}
+    } else {
+      const newEyecolor = '#'.concat(eyeColor)
+      return {eyeColor: newEyecolor}
+    }
   })
 }
