+++ src/lib/filters/get-data.js
@@ -1,5 +1,8 @@
-module.exports = function getData(datasource, key, variable) {
-  return datasource.map(person => ({
-    key: person.variable
-  }))
+module.exports = function getData(datasource, colors) { // get data from questionnaire file
+  return new Promise((resolve, reject) => {
+    const colors = datasource.map(person => ({
+      eyeColor: person.oogKleur
+    }))
+    resolve(colors)
+  })
 }
