+++ src/lib/filters/remove-whitespace.js
@@ -1,3 +1,6 @@
-module.exports = function removeWhitespace(data) {
-  return data.trim()
+module.exports = function removeWhitespace(data) { // remove excessive tabs at the start or end of a value
+  return data.map((person) => { // haal hier nog trimmed weg en return heel data.map
+    let trimmedEyeColor = person.eyeColor.trim()
+    return {eyeColor: trimmedEyeColor}
+  })
 }
